Data is now being pulled from the database to the fusioncharts object.
Added a switch case statement that feeds in data depending on the graph type.

Currently this only works for angular, but once the other configurations are done, this will be a fully functional script.

The titles aren't being pulled for some reason though, need to look into that, but data is being pulled.

Will be working to fix the title issue and implement other graph configurations.

GGWP.